chore(desktop): address the review on MCP UI app forwarding
Code-structure pass from the review on PR #96638; no behavior
changes.

- Name the "renders alone" concept: `rendersStandalone` in ChatThread
  wraps the three predicates, and `hasUiAppResult` moves to the
  mcp-apps feature, matching where `isPlanItem` and `isShowActionsItem`
  live.
- Move `omitNullCallToolResultFields` from `tool-meta.ts` to its own
  module, `shared/src/call-tool-result.ts`, so tool-meta keeps one
  subject (_meta handling).
- Document which failure each null-strip call site stops: the adapter
  strip is source hygiene for stored transcripts and McpAppsService
  events; `toCallToolResult` owns the schema-valid result an app
  receives.
- State in a comment why `keepMounted` on the group body is not the
  fix for the chart-exclusion rule.
- Clarify the lifecycle effect's contract with `sendResultOnce` in its
  deps, and the remount send's path through `sendWhenReady`.

Generated-By: PostHog Desktop
Task-Id: c63c7dce-f6d5-4ddd-a5fb-20f458bfa03a
